Changes: - added doc for all handlers - added doc for ratelimiter pakage - move ratelimiter to pkg folder - added example for ratelimiter pkg
